requirements.txt
streamlit
pandas
PyPDF2
pdfplumber
